feat: restructure role and permission management with new modules and DTOs
