+++ prisma/schema.prisma
@@ -62,14 +62,19 @@ model Role {
   deleted_at    DateTime?
   users         User[]
   permissions   RolePermission[]
+  createdAt     DateTime        @default(now())
+  updatedAt     DateTime        @updatedAt
 }
 
 model Permission {
   id            Int            @id @default(autoincrement())
   name          String          @unique
+  description   String?
   roles         RolePermission[]
   is_deleted    Boolean         @default(false)
   deleted_at    DateTime?
+  createdAt     DateTime        @default(now())
+  updatedAt     DateTime        @updatedAt
   @@index([name])
 }
 
@@ -80,7 +85,8 @@ model RolePermission {
 
   role          Role            @relation(fields: [roleId], references: [id])
   permission    Permission      @relation(fields: [permissionId], references: [id])
-
+  createdAt     DateTime        @default(now())
+  updatedAt     DateTime        @updatedAt
   @@unique([roleId, permissionId])
 }
