+++ src/system-permission/permission/permissions.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { PermissionController } from './permission.controller';
+import { PermissionsService } from './permissions.service';
+
+@Module({
+  controllers: [PermissionController],
+  providers: [PermissionsService],
+})
+export class PermissionsModule {}
