+++ src/system-permission/role/role.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { RoleController } from './role.controller';
+import { RoleService } from './role.service';
+
+@Module({
+  controllers: [RoleController],
+  providers: [RoleService],
+})
+export class RoleModule {}
